requirements.txt
numpy
scipy
matplotlib
pyyaml
tqdm
cdflib
opencv-python
einops
rich
onnxruntime-gpu==1.17.1 #requires cuda 12.2.0
transforms3d
ijson
pandas
poseutils

